+++ src/mkvtoolnix-gui/main_window/preferences_dialog.cpp
@@ -2,6 +2,7 @@
 
 #include <QFileDialog>
 #include <QIcon>
+#include <QInputDialog>
 #include <QItemSelectionModel>
 #include <QModelIndex>
 #include <QStandardItem>
@@ -92,6 +93,7 @@ PreferencesDialog::PreferencesDialog(QWidget *parent,
   setupEnableMuxingTracksByLanguage();
   setupMergeAddingAppendingFilesPolicy();
   setupMergeWarnMissingAudioTrack();
+  setupMergePredefinedTrackNames();
   setupTrackPropertiesLayout();
 
   // Info tool page
@@ -452,6 +454,10 @@ PreferencesDialog::setupConnections() {
 
   connect(ui->pbMDeriveTrackLanguageRevertCustomRegex,    &QPushButton::clicked,                                         this,                                 &PreferencesDialog::revertDeriveTrackLanguageFromFileNameRegex);
 
+  connect(ui->lwMPredefinedTrackNames,                    &QListWidget::itemSelectionChanged,                            this,                                 &PreferencesDialog::enablePredefinedTrackNameControls);
+  connect(ui->pbMAddPredefinedTrackName,                  &QPushButton::clicked,                                         this,                                 &PreferencesDialog::addPredefinedTrackName);
+  connect(ui->pbMRemovePredefinedTrackNames,              &QPushButton::clicked,                                         this,                                 &PreferencesDialog::removePredefinedTrackNames);
+
   connect(ui->cbGuiRemoveJobs,                            &QCheckBox::toggled,                                           ui->cbGuiJobRemovalPolicy,            &QComboBox::setEnabled);
   connect(ui->cbGuiRemoveOldJobs,                         &QCheckBox::toggled,                                           this,                                 &PreferencesDialog::adjustRemoveOldJobsControls);
   connect(ui->sbGuiRemoveOldJobsDays,                     static_cast<void (QSpinBox::*)(int)>(&QSpinBox::valueChanged), this,                                 &PreferencesDialog::adjustRemoveOldJobsControls);
@@ -622,6 +628,17 @@ PreferencesDialog::setupMergeWarnMissingAudioTrack() {
   });
 
   Util::fixComboBoxViewWidth(*ui->cbMWarnMissingAudioTrack);
+}
+
+void
+PreferencesDialog::setupMergePredefinedTrackNames() {
+  ui->lwMPredefinedTrackNames->addItems(Util::Settings::get().m_mergePredefinedTrackNames);
+  ui->pbMRemovePredefinedTrackNames->setEnabled(false);
+
+  for (int row = 0, numRows = ui->lwMPredefinedTrackNames->count(); row < numRows; ++row) {
+    auto item = ui->lwMPredefinedTrackNames->item(row);
+    item->setFlags(item->flags() | Qt::ItemIsEditable);
+  }
 }
 
 void
@@ -826,6 +843,14 @@ PreferencesDialog::save() {
   for (auto const &type : ui->tbMEnableMuxingTracksByType->selectedItemValues())
     m_cfg.m_enableMuxingTracksByTheseTypes << static_cast<Merge::TrackType>(type.toInt());
 
+  // Predefined track names:
+  m_cfg.m_mergePredefinedTrackNames.clear();
+  for (int row = 0, numRows = ui->lwMPredefinedTrackNames->count(); row < numRows; ++row) {
+    auto name = ui->lwMPredefinedTrackNames->item(row)->text();
+    if (!name.isEmpty())
+      m_cfg.m_mergePredefinedTrackNames << name;
+  }
+
   m_cfg.save();
 }
 
@@ -853,6 +878,28 @@ PreferencesDialog::enableOutputFileNameControls() {
   Util::enableWidgets(QList<QWidget *>{} << ui->gbDestinationDirectory   << ui->cbMUniqueOutputFileNames << ui->cbMAutoDestinationOnlyForVideoFiles, isChecked);
   Util::enableWidgets(QList<QWidget *>{} << ui->leMAutoSetFixedDirectory << ui->pbMBrowseAutoSetFixedDirectory, isChecked && fixedSelected);
   ui->leMAutoSetRelativeDirectory->setEnabled(isChecked && relativeSelected);
+}
+
+void
+PreferencesDialog::enablePredefinedTrackNameControls() {
+  ui->pbMRemovePredefinedTrackNames->setEnabled(!ui->lwMPredefinedTrackNames->selectedItems().isEmpty());
+}
+
+void
+PreferencesDialog::addPredefinedTrackName() {
+  auto newName = QInputDialog::getText(this, QY("Enter predefined track name"), QY("Please enter the new predefined track name."));
+  if (newName.isEmpty())
+    return;
+
+  auto newItem = new QListWidgetItem{newName};
+  newItem->setFlags(newItem->flags() | Qt::ItemIsEditable);
+  ui->lwMPredefinedTrackNames->addItem(newItem);
+}
+
+void
+PreferencesDialog::removePredefinedTrackNames() {
+  for (auto const &item : ui->lwMPredefinedTrackNames->selectedItems())
+    delete item;
 }
 
 void